+++ backend/packages/spend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/spend
 
+## 0.10.25
+
+### Patch Changes
+
+- Updated dependencies [704c99e]
+  - @cat-factory/contracts@0.46.0
+  - @cat-factory/kernel@0.47.2
+
 ## 0.10.24
 
 ### Patch Changes